headerLineExtractor initial implementation
